+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- [Swarm](resources/schema/swarm.graphql#L1) updated fields.
+
 ## [0.1.1] 2017-09-25
 
 ### Added